Add legend_sizes argument to layout().
This allows the size of legends to be modified. For example, on the TKL layout, many of the nav key's legends are too long to fit on a 1u key with the standard 6 font size, so the legend_sizes allows a change of this.

The array specifies differences from the default, so that if the default is adjusted the changed font sizes will be too. For example, +2 means a 2 unit larger legend, 0 means default and -1 means 1 unit smaller.